[Windows] Add check on service start to see if config needs to be updated.
Refactored config upgrade code from cmd/agent/app so it could be shared.
